+++ Changelog.md
@@ -5,6 +5,8 @@
 - Validate descriptor JSON file before deployment to k8s cluster
 - Upgrade ArgoCD to v2.5.5
 - Move Descriptor Validator to git-ops-command.sh
+- Upgrade nginx-ingress-controller to v1.5.1
+- Add base & region values.yaml files for Helm migration
 - Add multiple USER_BASE_DNs and BACKEND_IDs env vars 
 
 _Changes:_
@@ -12,6 +14,8 @@ _Changes:_
 - [X] PDO-4570 Validate descriptor JSON file before deployment to k8s cluster
 - [X] PDO-4575 Upgrade ArgoCD to v2.5.5
 - [X] PDO-4636 Move Descriptor Validator to git-ops-command.sh
+- [X] PDO-4698 Upgrade nginx-ingress-controller to v1.5.1
+- [X] PDO-4773 Update generate-cluster-state script to create base and region values.yaml files
 - [X] PDO-4818 Add multiple USER_BASE_DNs and BACKEND_IDs env vars 
 
 ### 1.17.0.0